Fix circular dependency in react-core
