+++ packages/react-core/src/core/providers/thirdweb-provider.tsx
@@ -1,4 +1,4 @@
-import { ThirdwebAuthProvider } from "../../evm/contexts/thirdweb-auth";
+import { ThirdwebAuthProvider } from "../../evm/contexts/ThirdwebAuthProvider";
 import { useUpdateChainsWithApiKeys } from "../../evm/hooks/chain-hooks";
 import { ThirdwebSDKProvider } from "../../evm/providers/thirdweb-sdk-provider";
 import { ThirdwebSDKProviderProps } from "../../evm/providers/types";
