+++ packages/react-core/src/evm/contexts/thirdweb-auth.tsx
@@ -1,12 +1,4 @@
-import React, {
-  PropsWithChildren,
-  createContext,
-  useContext,
-  useEffect,
-  useMemo,
-} from "react";
-import { useLogout, useUser } from "../hooks/auth";
-import { useWallet } from "../../core/hooks/wallet-hooks";
+import { createContext, useContext } from "react";
 
 export interface ISecureStorage {
   getItem(key: string): Promise<string | null>;
@@ -44,66 +36,12 @@ export interface ThirdwebAuthConfig {
   secureStorage?: ISecureStorage;
 }
 
-interface ThirdwebAuthContext extends ThirdwebAuthConfig {}
+export interface ThirdwebAuthContext extends ThirdwebAuthConfig {}
 
-const ThirdwebAuthContext = createContext<ThirdwebAuthContext | undefined>(
-  undefined,
-);
-
-export const ThirdwebAuthProvider: React.FC<
-  PropsWithChildren<{ value?: ThirdwebAuthConfig }>
-> = ({ value, children }) => {
-  // Remove trailing slash from URL if present
-  const authContext = useMemo(() => {
-    if (!value) {
-      return undefined;
-    }
-
-    const context: ThirdwebAuthContext = {
-      ...value,
-      authUrl: value.authUrl?.replace(/\/$/, ""),
-    };
-
-    return context;
-  }, [value]);
-
-  return (
-    <ThirdwebAuthContext.Provider value={authContext}>
-      {children}
-      <DisconnectOnAccountSwitch />
-    </ThirdwebAuthContext.Provider>
-  );
-};
+export const ThirdwebAuthContext = createContext<
+  ThirdwebAuthContext | undefined
+>(undefined);
 
 export function useThirdwebAuthContext() {
   return useContext(ThirdwebAuthContext);
-}
-
-function DisconnectOnAccountSwitch() {
-  const { logout } = useLogout();
-  const wallet = useWallet();
-  const { isLoggedIn } = useUser();
-
-  useEffect(() => {
-    const handleChange = (data: { address?: string; chainId?: number }) => {
-      // if the user changes their account, logout
-      if (data.address) {
-        logout();
-      }
-    };
-
-    const shouldAddListener = wallet && isLoggedIn;
-
-    if (shouldAddListener) {
-      wallet.addListener("change", handleChange);
-    }
-
-    return () => {
-      if (shouldAddListener) {
-        wallet.removeListener("change", handleChange);
-      }
-    };
-  }, [wallet, logout, isLoggedIn]);
-
-  return null;
 }
